[Refactor] Split scan task queue implementation by queue type
