+++ be/src/exec/pipeline/lookup_operator.h
@@ -19,7 +19,7 @@
 #include "common/status.h"
 #include "exec/pipeline/operator_factory.h"
 #include "exec/pipeline/source_operator.h"
-#include "exec/workgroup/scan_task_queue.h"
+#include "exec/workgroup/scan_task.h"
 #include "exec/workgroup/work_group_fwd.h"
 #include "runtime/descriptors_fwd.h"
 #include "runtime/lookup_stream_mgr.h"
