+++ be/src/exec/pipeline/sink/file_sink_operator.cpp
@@ -23,7 +23,6 @@
 #include "exec/pipeline/query_context.h"
 #include "exec/pipeline/sink/sink_io_buffer.h"
 #include "exec/workgroup/scan_executor.h"
-#include "exec/workgroup/scan_task_queue.h"
 #include "exprs/expr.h"
 #include "exprs/expr_executor.h"
 #include "exprs/expr_factory.h"
