+++ be/src/exec/spill/executor.h
@@ -24,6 +24,7 @@
 #include "exec/pipeline/query_context.h"
 #include "exec/workgroup/pipeline_executor_set.h"
 #include "exec/workgroup/scan_executor.h"
+#include "exec/workgroup/scan_task.h"
 #include "exec/workgroup/scan_task_queue.h"
 #include "exec/workgroup/work_group_fwd.h"
 #include "gen_cpp/Types_types.h"
