+++ be/src/runtime/exec_env.cpp
@@ -64,7 +64,6 @@
 #include "exec/spill/spill_metrics.h"
 #include "exec/workgroup/pipeline_executor_set.h"
 #include "exec/workgroup/scan_executor.h"
-#include "exec/workgroup/scan_task_queue.h"
 #include "exec/workgroup/work_group.h"
 #include "exec/workgroup/work_group_manager.h"
 #include "fs/fs_s3.h"
