+++ be/src/udf/java/utils.cpp
@@ -21,7 +21,7 @@
 #include "common/compiler_util.h"
 #include "common/thread/priority_thread_pool.hpp"
 #include "exec/workgroup/scan_executor.h"
-#include "exec/workgroup/scan_task_queue.h"
+#include "exec/workgroup/scan_task.h"
 #include "exec/workgroup/work_group.h"
 #include "exec/workgroup/work_group_manager.h"
 #include "runtime/current_thread.h"
