+++ build-support/exec_env_singleton_allowlist.txt
@@ -15,7 +15,7 @@ src/exec/pipeline/query_context_manager.cpp
 src/exec/schema_scanner/schema_be_cloud_native_compactions_scanner.cpp
 src/exec/schema_scanner/schema_be_tablets_scanner.cpp
 src/exec/workgroup/lock_free_work_group_scan_task_queue.cpp
-src/exec/workgroup/scan_task_queue.cpp
+src/exec/workgroup/work_group_scan_task_queue.cpp
 src/exec/write_combined_txn_log.cpp
 src/fs/fs_broker.cpp
 src/service/service_be/config_update_hooks.cpp
